+++ src/main/java/edu/uminho/biosynth/core/data/io/dao/biodb/bigg/DefaulBiggReactionParser.java
@@ -51,7 +51,7 @@ public class DefaulBiggReactionParser {
 		rxn.setEnzyme(values[5]);
 		rxn.setOrientation(values[6].equals("Reversible") ? 0 : 1);
 		rxn.setTranslocation(values[7].equals("N") ? false:true);
-		rxn.setId(Integer.parseInt(values[8]));
+		rxn.setId(Long.parseLong(values[8]));
 		for (String modelIntValue : values[9].split(",")) {
 			String modelId = convertToModelCrossReference(Integer.parseInt(modelIntValue));
 			if (modelId != null) {
